updated main and samples readme  (#135)
